refactor(backend): extract exporter service and resolver
