+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "crates/migrations",
     "crates/models",
     "crates/providers",
+    "crates/resolvers",
     "crates/services",
     "crates/traits",
     "crates/utils",
@@ -66,7 +67,6 @@ serde = { version = "=1.0.204", features = ["derive"] }
 serde_json = "=1.0.120"
 serde_with = { version = "=3.9.0", features = ["chrono_0_4"] }
 strum = { version = "=0.26.2", features = ["derive"] }
-struson = { version = "=0.5.0", features = ["serde"] }
 # FIXME: Upgrade once https://github.com/seanmonstar/reqwest/pull/1620 is merged
 reqwest = { git = "https://github.com/thomasqueirozb/reqwest", branch = "base_url", features = [
     "json",
