+++ crates/services/exporter/Cargo.toml
@@ -11,6 +11,7 @@ chrono = { workspace = true }
 common_models = { path = "../../models/common" }
 common_utils = { path = "../../utils/common" }
 database_models = { path = "../../models/database" }
+database_utils = { path = "../../utils/database" }
 config = { path = "../../config" }
 file_storage_service = { path = "../file_storage" }
 fitness_models = { path = "../../models/fitness" }
@@ -20,7 +21,6 @@ nanoid = { workspace = true }
 reqwest = { workspace = true }
 sea-orm = { workspace = true }
 serde_json = { workspace = true }
-struson = { workspace = true }
+struson = { version = "=0.5.0", features = ["serde"] }
 tokio = { workspace = true }
 tokio-util = { version = "=0.7.11", features = ["codec"] }
-database_utils = { path = "../../utils/database" }
